requirements.txt
pygame==2.1.2
pylint==2.12.2
pytest==6.2.5
python-dotenv==0.19.2
PySimpleGUI==4.55.1
spotipy==2.19.0

